requirements.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.6.5
django-mongodb-engine==0.5.2
django-toolbelt==0.0.1
djangorestframework==3.1.0
djangotoolbox==1.6.2
gunicorn==19.0.0
httpie==0.9.2
httplib2==0.9
psycopg2==2.5.3
Pygments==2.0.2
pymongo==2.8
pytz==2014.10
requests==2.3.0
six==1.9.0
static3==0.5.1
twilio==3.7.3
